+++ yash-env/CHANGELOG.md
@@ -18,6 +18,7 @@ A _private dependency_ is used internally and not visible to downstream users.
       `yash_syntax::source`.
 - `parser` module
     - `Config`: Configuration for the parser.
+    - `IsKeyword`: Function that checks if a string is a reserved word.
     - `is_name`: Function that checks if a string is a valid variable name.
 - `alias` module
     - This module re-exports `Alias`, `AliasSet`, `Glossary`, and `HashEntry`